lb-csi: add build_id information
in order to be able to trace back the lb-csi-plugin binary and the
docker image we need to inject this information.

build_id is a env variables generated by the jenkins CI.
When we build on the release-job we will give the correct build-id and
will be able to trace back the image and when it was compiled.

issue: LBM1-13390