build: Use official busybox unless MISC_DOCKER_REGISTRY is defined
Currently, there is an asymmetry between the lb-csi-plugin, the busybox
and the sidecar images in the Makefile.
Sidecars have their own registry prefix which default to public registry and this is what makes them "work out of the box".
The lb-csi-plugin is pushed as part of the build process if a registry is specified.
The busybox image is not pushed to the private registry but the prefix is shared with the lb-csi-plugin docker-registry which breaks the flow of "build/tag/push".

In order to fix this, simply use a different docker-registry prefix to
the busybox image so it will default to the public dockerhub if not
specified, which will make it work out of the box.

For customers who cannot access public registries, the
MISC_DOCKER_REGISTRY environment variable can be set just like the
SIDECAR_DOCKER_REGISTRY to specify the location of misc packages such as
busybox.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@email.com>