controller: align error values according to spec
running csi-sanity we see that:

DeleteVolume:
* should fail when no volume id is provided.
* should succeed when an invalid volume id is used.

This coresponds with the claim in the spec:

"This operation MUST be idempotent. If a volume corresponding to the
specified volume_id does not exist or the artifacts associated with the
volume do not exist anymore, the Plugin MUST reply 0 OK."

We introduce error types in order to improve error reporing on the
controller level.