controller: expose CREATE_DELETE_SNAPSHOT support
This patch comes after we have full support for creating and deleting
snapshots and also creating a volume from a snapshots content source.

LIST_SNAPSHOTS is unneeded at this point because if the csi driver
doesn't support this, then k8s automatically assumes it's own snapshot
DB entries are valid.

Signed-off-by: Sagi Grimberg <sagi@lightbitslabs.com>