node: NodePublishVolume add readonly support
The NodePublishVolumeRequest contains a readonly option:
  // Indicates SP MUST publish the volume in readonly mode.
  // This field is REQUIRED.
  bool readonly = 6;

When readonly is set in the request, the SP (Storage Provider) is
expected to "publish" the volume in readonly mode. This means that the
volume published into the container is in readonly mode, and has nothing
to do with the actual state of the volume in LightOS.
Publishing a volume as readonly is done by simply adding the "ro" flag
to the mount options.

It is therefore a BUG that currently the CSI plugin refuses to publish
volumes as readonly - which is also evident as the only e2e tests which
fail are the ones which request to publish volumes as readonly.

Fix this by adding "ro" to the mountOptions in NodePublishVolume.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>