driver: poll for secret on volume mount
In some deployments where we don't want to define specific secrets for
specific APIs we might use a secret and attach it to the controller and
node pods.
Data in the secret will hold the JWT we would use for all api calls
against LightOS.

since env vars can't change dynamically we change the way we expose the
JWT to the plugin from env var to volume mounts.