node: Fix device path detection
Running our csi plugin on a node with
local NVMe devices, now these devices happen to be rather
old and present unexpected identification to the host (not
uuid, not nguid, not eui) so the host builds an identification
based on the device serial, model, vendor and nsid.

In addition, the nvme core driver also presents to sysfs only
identifiers based on their exposure.

There are two issues found:
1. We don't filter partitions, which cause sysfs attribute read
   error we don't expect (so we filter partition device nodes)
2. We tried to get the device identifier from nguid sysfs attribute
   which is not presented by the kernel if the device doesn't expose
   it. So we get the device identifier from /sys/block/nvmeXnY/wwid which
   is the world-wide identifier and strip the identifier prefix.

In addition, given that we already found some issues on this area
instead of failing getting the device path based on UUID, we simply
log errors but ignore the error to allow at least to complete
the search. We log the fact that we were not able to get the device
and will simply continue to loop until halted.

And while we're at it, add some debug logging in the failure path.

Signed-off-by: Sagi Grimberg <sagi@lightbitslabs.com>