deployment: Update snapshot-controller, csi-snapshotter and CRDs to v4.0.0
When running e2e tests on new kubernetes clusters, the snapshotable test
suite fails since it tries to access snapshot.storage.k8s.io/v1 api,
which is not available in the CRDs we have at the moment.

Support for v1 has been added in the external-snapshotter after v3.0.3
(we have v3.0.0) - and it is backward compatible (v1beta1 is kept), so
upgrade to the latest offical release which is v4.0.0.

While at it, update the examples to use v1 API which proves that it
works, so e2e tests should be fine now.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>