test: e2e: test.sh: Add SupportedSizeRange to DriverInfo in testdriver.yaml
The e2e snapshot tests is requesting 1MiB PVC, but LightOS has 1GiB
minimum size so the resulting PVC is 1GiB.
The test then creates a snapshot, which is also 1GiB, then requests to
create a new 1MiB PVC from that snapshot. This fails at the
csi-provisioner due to the requested volume being too small to contain
the snapshot.

The solution is to use the SupportedSizeRange configuration in the e2e
tests and set the minimum to 1GiB.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>