client: always set snapshot creation timestamp
The CSI specification says that the creation_time field is required, but
LightOS response contains a nil value as long as the snapshot state is
"creating".

The external-snapshotter treats this as error, which eventually is
causing e2e snapshottable tests to fail.

Workaround that without patching the external-snapshotter by setting an
empty timestamp if the response from LightOS came out with a nil
creationTime.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>