controller: refactor doCreateVolume - add validateVolume
doCreateVolume has two code paths:
1. The first time the CO calls ControllerCreateVolume, the CSI driver
starts a blocking call to LightOS to create the volume.
2. If CO gave up on the ControllerCreateVolume first call, LightOS might
have still been in the process of creating the volume. On consecutive
calls from the CO, the CSI driver will first check if a volume with the
given name exists.

Flow 2 has volume validation in that the volume we get from LightOS is
compared to the CO request.
Flow 1 has no volume validation.

This is a bug in the CSI driver - and it makes debugging issues harder.
For example, "LBM1-15016 - create volume from snapshot does not return
sourceSnapshotID on v2" was hard to reproduce because it sometimes
failed and sometimes passed, due to the fact we don't do the volume
validation in Flow 1.

Fix this by introducing a new validateVolume method which is always
called, and refactor doCreateVolume accordingly so now both code paths
call validateVolume.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>