lb-csi-workload-examples: create helm chart for examples
the examples we provide today are very specific and were appropriate in
the past when we had 2-3 examples which where static.

now we have 6 or more workload examples and we want to port it to helm
for customers to easily configure them on site.

Also it will provide a more real-world example for customers who
probably use helm in they day-to-day deployments on how to integrate our
storage with their k8s applications.

This will also help us in validation when we can programmatically modify
the examples to the running cluster and verify they are working.

We will be able to change the following:
1. storageClass parameters
2. jwt in secret
3. mgmt-endpoint
4. namespace
and much more...

We will provide the following workload examples:

1. storageclass -- deploy a SC, Secret, VolumeSnapshotClass.
2. block -- deploy PVC and POD consuming it
3. filesystem -- deploy PVC and POD consuming it
4. StatefulSet -- deploy StatefulSet on all nodes that use use a dynamic
provisioned PVC.
5. snaps -- comprised of multiple scenarios:
   1. create snap from PVC
   2. create PVC from snap taken in fase 1 and a POD using created PVC
   3. create PVC from PVC taken from step 2 and a POD using created PVC