Makefile: dont allow to gen bundle if docker_registry not set
The docker_registry is a mandatory field cause we use it to generate the
manifests.
It must have a sane value otherwize we would create a bundle that is not
usable.
In lab tests we will call Makefile.lb which will set DOCKER_REGISTRY to
lbdocker:5000. on package-csi.sh we will invoke it with
DOCKER_REGISTRY=cloudsmith/path to export it to customers.