drop busybox dependency
today we have an init-container that is requiring a busybox image
this image is fixed and most of the time needs to be added to the
private registry of the cusomter when deploying.
we also need to provide it in cloud-smith which we don't today and tell
the customer to bring it's own.

This creates another freqtion point that is unnceccery.
We can use our own lb-csi-plugin image for this task (loading the nvme
module) without requiring additional image (no matter how small).