go.mod: Remove dependecy on k8s.io/kubernetes
While I was trying to bump up the dependecy in go mod for
k8s.io/kubernetes I noticed we our go mod fails, and I could not
add a new depedency on volume subpackage.

There are some client packages like client-go and k8s.io/utils that are
meant to be imported and are simple to do.

There is an effort to extract common code to client packages so they can
be imported easily. Some example libraries are:
* k8s.io/utils/exec
* k8s.io/mount-utils
that enable to use the mount/resizefs functions.

all packages under k8s.io/kubernetes are not meant to be imported but
sometimes we must, cause there are useful stuff there.

See this post about the problem:
https://suraj.io/post/2021/05/k8s-import/

I came across this comment:
https://github.com/kubernetes/kubernetes/issues/79384#issuecomment-505627280
recomending using replace

Changes applied:
* use mount-utils package instead of k8s.io/kubernetes/pkg/utils/mount
* use mount-utils package instead of k8s.io/kubernetes/pkg/util/resizefs
* upgrade k8s.io/kubernetes dependency to 1.20
* upgrade csi spec to 1.3.0
* upgrade go.mod to golang 1.16 and update vendor folder
* scripts/download-deps.sh: add script to upgrade k8s deps easily

issue: LBM1-17861