Makefile: fix external build flow
We have 3 flows of building images:
1. CI with unique JOB_NAME/BUILD_ID variables.
2. Internal tests with access to dockerize and version.py using
Makefile.lb
3. external clients with no access to our CI and our dockerize image.

In the third scenario we need to generate unique image TAG for the
image. This image tag will be generated by git short description.

An external user running build will be able to use this auto-generated
hash and will be able to override it using BUILD_HASH variable.