fix builder image docker gid collision.
When building the image on some hosts we get the following:

addgroup: gid '999' in use
The command '/bin/sh -c addgroup -g $DOCKER_GID docker' returned a
non-zero code: 1

The reason is that alpine images have user named ping which we don't use
that is using 999 GID.

On some distros and docker version, docker GID is also 999 and that
will cause a collition.

So we check if that GID is taken in the image when building and removing
it if so.

issue: LBM1-18677