lightos-patcher: add script to automate LightOS cluster resize
Today the csi-plugin will embed in the ResourceID the mgmt-endpoint
information.
When we resize the cluster this information is not updated and is
immutable via the API.

This script will force update on this field by using the replace API
that affectivly will delete and recreated the resource with the updated
field.

We provide an update on the storage class as well as all the volumes
that were created referencing this storage-class.

This script should be idempotent.

issue: LBM1-18839