test/e2e/test.sh: return real errorcode from trap
since we used trap we didn't proprgate the real return-code from the
test.
following this answer:
https://unix.stackexchange.com/a/235586
added the trap return code to make sure the service will fail in case a
test had failed.
remove the tee to the log file since we fail to get the correct error
code of the test.e2e binary

add compare_semantic version method to check the cluster version.
we need to make sure if we run with >=1.20 to use the snapshot/v1 CRD
if we run below 1.20 we need to run with snapshot/v1beta1 CRD

Also improve the focus and skip flags to the test to make it more
readable.