controller: refactor DeleteSnapshot(), fix retvals
LB "clones" work though an intermediate snapshot. this adds a new,
internal call path to DeleteSnapshot() invocations, in addition to
the regular gRPC service one. the patchset that introduced clones
feature called DeleteSnapshot() directly, which wreaks havoc with
the logs (wrong output), and errors passing.

split DeleteSnapshot() into a core internal impl that actually deletes
as snap, and a gRPC-ready wrapper that sets everything up for it. the
CreateVolume() flow already has all of this stuff set up. the same fate
will before at least CreateSnapshot() later on in this patchset, BTW.

while at it, fix the retval on bad `volume_id` values: a very literal
interpretation of the CSI spec requires they be reported as NotFound,
and 3rd party test frameworks require it to be so, even though as far
as LB CSI plugin is concerned, not all VIDs are created equal...

plus cleanups, formatting, lints, etc.