WIP: CreateVolume() rewrite
fix at least some of the issues that have accumulated since the
introduction of block vol support, snapshots, clones, MT, etc...

NOTE: this patch does not alter the "intermediate snapshots" design
used when clones were originally introduced. it's very problematic, but
fixing it will be left as an exercise for a future date. see comments
int the code for details.

on clone (vol from snap/vol) flows:
* add checks that the source is actually compatible with the requested
  volume params: replica count, compression, capacity.
* make these checks transitive for vol clone via intermediate snap.
* check the source explicitly in order to return the correct error
  codes to satisfy the CSI spec (and 3rd party test frameworks!).
* make an actual effort to delete the intermediate snaps. it's still not
  watertight, obviously, but at least now the throwaway snaps are no
  longer liberally leaked as they used to be.
* fix authZ by actually passing the bearer tokens to be used by the
  `lbgrpc` client on the clone from vol flow.
* return CSI spec compliant errors on all resource ID parsing cases
  to ingratiate the LB CSI plugin with the 3rd party test frameworks.
* return CSI spec compliant errors (AFAICT) on intermediate steps'
  failures to give the process a higher chance to complete in the face
  of the substantial time some snaps-related ops take in LB.
* invoke internal versions of CreateSnapshot()/DeleteSnapshot()
  introduced earlier in the patchset to fix the completely bogus traces.

on existing volume found (retry or admin command) flows:
* check that the volume is actually based on the requested content
  source, directly for snaps, transiently through the snap back to
  source vol for clones.
* perform CSI spec compliant capacity range check against the source.
* undo validateVolume() damage: claims in "controller: refactor
  doCreateVolume - add validateVolume" notwithstanding, both flows WERE
  always checked. that patch just repeated the same checks twice,
  without fixing the original problem, because when snapshots support
  was added, a corresponding source ID check clause wasn't added to the
  checks list (and it remained disabled until a previous commit in this
  patchset fixed it).

this patch is NOT exhaustive, and even setting the TODOs already in the
code aside, there's still a bunch of checks that need to go in in order
to be compliant as it is, and more to make the interaction of additional
features (e.g. client-side encryption) with the existing ones more safe.
i still hope to beef this one up before merging...

issue: LBM1-19398
issue: LBM1-19400