lbgrpc: test: add TestClone()
leaf test to check a basic clone (i.e. a volume that is based on another
volume by means of an intermediate snapshot, as LB volumes can only be
based on snapshots) and a clone of a clone creation/deletion.

it assumes that basic snapshot and volume tests pass, instead of
rechecking everything from scratch.