test/e2e/test.sh: add skip list for each k8s version
There is a bug in e2e test older then v2.22.3 for preprovisioned
snapshot tests.

the issue is that they are never deleted because the test does not
annotate them with the secret info.

the following patch in the upstream fix it:

commit 13d852c73dc7f1b78a6c5933987bea7ceb721d18
Author: Mauricio Poppe <mauriciopoppe@google.com>
Date:   Thu Sep 30 23:43:21 2021 +0000

    Copy VolumeSnapshotContent annotations in snapshottable.go test

In our case patching it would be cumbersome cause the code was rewritten
twice and cherry-pick is not strait forward.

This patch will add a skip-list for every k8s version and we choose to
skip these 2 test cases that fail.

issue: LBM1-19772