Makefile: pass in GIT_VER from outside image-builder
since we use git-repo our .git folder points to the
../../.repo/projects/lb-csi.git folder.

we don't mount this folder in the image-builder container, so we have no
access to the .git folder from inside the image-builder.

This patch will calculate the GIT_VER outside the image builder and
will not need to recalculate in it inside the image-builder cause it is
already defined.