Makefile: standardize image labels
looked for a standard definition of a set of labels that should be
assigned to an image.
there is this initiative - http://label-schema.org/rc1/ - but it was
superseeded by the following:

https://github.com/opencontainers/image-spec/blob/main/annotations.md#pre-defined-annotation-keys

This is not mandetory but a common practice to label containers.

figured we should align with such a standard.