driver: Add support for luks host-encryption
When a StorageClass with host-encryption set to 'enabled', we are
going to set up a luks2 encrypted device overlay on top of the
host nvme device.

For host-encryption, we expect and require the relevant csi invocations
to be fed with a host-encryption-passphrase derived from an appropriate
secret.

This commit also introduces a luks library used for luks device
management.

We enrich the lbResourceID with an optional 'hostcrypto:luks2' string
that will be set for host-encrypted volumes and snapshots. This is
used as the indication that the resource (volume/snapshot) is or
should be host-encrypted with a luks2 overlay device.

The relevant flows where this feature is touching is:
1. controller: mainly indicate and check that the resource either should
               be host encrypted, or verify that the resource requirements
	       match host-encryption (volumes and/or snapshots).
2. csi_md: assemble and parse lbResourceID with an optional hostcrypto
	   attribute.
3. node:
	- NodeStageVolume: where we need to open and optionally format a luks2
	  device (for fs mounts) if host-encryption is requested.
	- NodeUnstageVolume: close the luks2 device
	- nodePublishVolumeForBlock: set the source path to the luks2 device
	  rather than the nvme namesapce device.
	- NodeExpandVolume: proactively expand the luks2 device.
4. throughout: add host-encryption attribute logging when relevant.

One more thing to note:
When a host-encrypted volume created from a non-encrypted snapshot, we catch it
and validate it correctly. However when a host-encrypted volume is created from
a host-encrypted snapshot but with different passphrases, we will fail to open
the luks device as we cannot store the passphrase in the lbResourceID, hence
when we fail to open the luks device, we log that this may have happened with
a generic statement of: "(double-check that passphrases match)" as the output
from luksOpen indicate a passphrase related error.

based on work from: Stefan Majer <Stefan.Majer@f-i-ts.de>

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>