csi_md.go: expose AccessMode_MULTI_NODE_MULTI_WRITER only if RWX is set
Since this feature can lead to data-corruption, we don't want to enable
it by default, hence we do enable it and publish this capability only if
this flag (LB_CSI_RWX) was configured to 'true' and the accessMode is
Block.

We don't allow volumeMode==Filesystem, since the driver does mkfs on
stage/publish steps which may cause data corruptions if not syncronized.

Reveresed the order of the checkCapability in order to pass volumeMode
to supportedAccessModes method.

issue: LBM1-12623