luks: Add some dm-crypt performance flags
There are 4 flags that are supported for performance enhancements
(and we care about performance), see `man cryptsetup` for more details:
1. --perf-same_cpu_crypt
2. --perf-submit_from_crypt_cpus
3. --perf-no_read_workqueue
4. --perf-no_write_workqueue

It is fine to blindly add them, because cryptsetup will hide the
unsupported failures from us by retrying without them. The only
side effect is that it will slightly spam the log with dm-crypt
error messages with the failed attempts.

Signed-off-by: Sagi Grimberg <sagi@grimberg.me>