enable luks configuration from file
today we have static sane defaults in our code to configure client side enc.
This does not give us the freedom to support different configurations
depending on hardware requirements.

by providing a configuration file to all luks related config we
enable a way to override values on a per node basis.

This means that we get an extra level of freedom to provide custom
config for each node if we so choose.