build(image-builder): rename and uniquly version builder image
till now the image builder name was staticlly version to 0.0.1
this may have caused changes to the image to not be updated
in case the image was already present on the machine.

now we version the image with a unique hash on every Dockerfile
change, meaning each change will be detected.