build(Makefile): fix image url calculation
We want the default image path to be:

${DOCKER_REGISTRY}/${ORG}/${NAME}:${VERSION}

But we want to allow a user to override the different parts.
this means that:

Default values are:

DOCKER_REGISTRY==docker.lightbits.com
ORG=lightos-csi
NAME=lb-nvme-discovery-client

A user can provide a different REGISTRY or a REGISTRY/ORG
and the image name will be override accordingly.