fix: explicitly specify fsType in StorageClass example template
today we give helm templates as examples for generating StorageClasses for los-csi.
this exmaple is missing a key flag csi.storage.k8s.io/fstype when we specify
the fsType flag, so k8s will not configure fsGroup settings if this flag is
missing from StorageClass.

actually there are 2 issues: first - the default value is empty (assuming ext4)
but it will result in us not specifiying the csi.storage.k8s.io/fstype param.
second - there was a typo in the fstype/fsType field (uppercase T) causing
the template never to render this field.

issue: LBM1-20579